+++ src/main/java/io/github/natank25/epitechutils/files/EpitechTemplates.java
@@ -6,7 +6,13 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiDirectory;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiManager;
+import com.jetbrains.cidr.cpp.makefile.MakefileUtil;
+import com.jetbrains.cidr.cpp.makefile.actions.MakefileRefreshProjectAction;
+import com.jetbrains.cidr.cpp.makefile.project.MakefileAutoImportAware;
+import com.jetbrains.cidr.cpp.makefile.settings.MakefileProjectSettings;
+import com.jetbrains.cidr.cpp.makefile.settings.MakefileSettings;
 import io.github.natank25.epitechutils.actions.EpitechNewFileAction;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -17,7 +23,8 @@ import java.util.Properties;
 public class EpitechTemplates {
 	public static final String C_FILE = "Epitech C File.c";
 	public static final String HEADER_FILE = "Epitech Header File.h";
-	public static final String MAKEFILE = "Makefile.mk";
+	public static final String MAKEFILE = "Epitech Makefile.mk";
+	public static final String GITIGNORE = ".gitignore";
 	
 	public static @NotNull PsiElement createFromTemplate(Project project, String fileName, PsiDirectory dir, Templates template, @Nullable Properties properties){
 		Properties defaultProperties = EpitechNewFileAction.createProperties(project, fileName);
@@ -25,7 +32,7 @@ public class EpitechTemplates {
 			defaultProperties.putAll(properties);
 		try {
 			return FileTemplateUtil.createFromTemplate(
-					FileTemplateManager.getInstance(project).getTemplate(template.toString()),
+					FileTemplateManager.getInstance(project).getInternalTemplate(template.toString()),
 					fileName,
 					properties,
 					dir);
@@ -50,16 +57,25 @@ public class EpitechTemplates {
 		return createFromTemplate(project, fileName, PsiManager.getInstance(project).findDirectory(dir), Templates.HEADER_FILE);
 	}
 	
+	public static @NotNull PsiElement createGitignoreFileFromTemplate(Project project, VirtualFile dir, String binary_name){
+		Properties gitignorProperties = new Properties();
+		gitignorProperties.setProperty("BIN_NAME", binary_name);
+		return createFromTemplate(project, ".gitignore", PsiManager.getInstance(project).findDirectory(dir), Templates.GITIGNORE, gitignorProperties);
+	}
+	
 	public static @NotNull PsiElement createMakefileFileFromTemplate(Project project, VirtualFile dir, String binary_name){
 		Properties makefileProperties = new Properties();
 		makefileProperties.setProperty("BIN_NAME", binary_name);
-		return createFromTemplate(project, "Makefile", PsiManager.getInstance(project).findDirectory(dir), Templates.MAKEFILE, makefileProperties);
+		PsiElement file = createFromTemplate(project, "Makefile", PsiManager.getInstance(project).findDirectory(dir), Templates.MAKEFILE, makefileProperties);
+		((PsiFile) file).setName("Makefile");
+		return file;
 	}
 	
 	public enum Templates{
 		C_FILE(EpitechTemplates.C_FILE),
 		HEADER_FILE(EpitechTemplates.HEADER_FILE),
-		MAKEFILE(EpitechTemplates.MAKEFILE);
+		MAKEFILE(EpitechTemplates.MAKEFILE),
+		GITIGNORE(EpitechTemplates.GITIGNORE);
 		
 		private final String filename;