+++ src/main/java/io/github/natank25/epitechutils/icons/EpitechUtilsIcons.java
@@ -1,0 +1,9 @@
+package io.github.natank25.epitechutils.icons;
+
+import com.intellij.openapi.util.IconLoader;
+
+import javax.swing.*;
+
+public class EpitechUtilsIcons {
+	public static final Icon EpitechIcon_150x150 = IconLoader.getIcon("/icons/EpitechLogo/epitech-logo-150x150.svg", EpitechUtilsIcons.class);
+}